+++ example/main.dart
@@ -28,7 +28,7 @@ main() {
   var personClassMirror = reflectType(Person);
   // and then constructs a new person using a map with the
   // needed parameters for the constructor
-  var p1 = personClassMirror.constructors[''].call({'id': 1, 'name': 'person 1'});
+  var p1 = personClassMirror.constructors['']([], {'id': 1, 'name': 'person 1'});
   // Get the list of DeclarationMirror corresponding to the fields of Person class
   var p1Fields = personClassMirror.fields;
 
@@ -39,7 +39,7 @@ main() {
 
   // Gets the CarClassMirror and constructs a new car using the default constructor
   // passing a map containing the required parameters
-  Car car1 = reflectType(Car).constructors[''].call({'id': 1, 'engine': 'v8'});
+  Car car1 = reflectType(Car).constructors['']([1, 'v8']);
   /* prints:
       car1:
         id: 1
@@ -57,9 +57,9 @@ main() {
   print(methods.keys); // prints: 'someFunction'
   print(methods['someMethod'].returnType); // prints: String
   print(methods['someMethod'].annotations); // prints: [Instance of '_MyOtherAnnotation']
-  print(methods['someMethod'].parameters); // prints: {p1: Instance of 'DeclarationMirror'}
-  print(methods['someMethod'].parameters['someParameter'].annotations); // prints: [Instance of '_MyOtherAnnotation']
-  print(methods['someMethod'].parameters['someParameter'].type); // prints: int
+  print(methods['someMethod'].positionalParameters); // prints: {p1: Instance of 'DeclarationMirror'}
+  print(methods['someMethod'].positionalParameters[0].annotations); // prints: [Instance of '_MyOtherAnnotation']
+  print(methods['someMethod'].positionalParameters[0].type); // prints: int
 
   print('\n--------------------------');
   print('reflecting "someFunction"');
@@ -68,10 +68,10 @@ main() {
   print(sfMirror.name); // prints: '(someMethod)'
   print(sfMirror.returnType); // prints: dynamic
   print(sfMirror.annotations); // prints: [Instance of '_MyOtherAnnotation']
-  print(sfMirror.parameters); // prints: {someParameter: Instance of 'DeclarationMirror'}
-  print(sfMirror.parameters['p1'].annotations); // prints: [Instance of '_MyOtherAnnotation']
-  print(sfMirror.parameters['p1'].type); // prints: String
-  print(sfMirror.parameters.values.elementAt(0).name);
-  print(sfMirror.parameters.values.elementAt(1).name);
-  print(sfMirror.parameters.values.elementAt(2).name);
+  print(sfMirror.positionalParameters); // prints: {someParameter: Instance of 'DeclarationMirror'}
+  print(sfMirror.positionalParameters[0].annotations); // prints: [Instance of '_MyOtherAnnotation']
+  print(sfMirror.positionalParameters[0].type); // prints: String
+  print(sfMirror.positionalParameters[0].name);
+  print(sfMirror.positionalParameters[1].name);
+  print(sfMirror.positionalParameters[2].name);
 }