+++ test/models/course.g.dart
@@ -6,7 +6,7 @@ part of built_mirrors.test.models.course;
 // Generator: MirrorsGenerator
 // **************************************************************************
 
-_Course__Constructor(params) => new Course();
+_Course__Constructor([positionalParams, namedParams]) => new Course();
 
 const $$Course_fields_id = const DeclarationMirror(type: int);
 const $$Course_fields_name = const DeclarationMirror(type: String);
@@ -14,7 +14,7 @@ const $$Course_fields_credits = const DeclarationMirror(type: num);
 
 const CourseClassMirror =
     const ClassMirror(name: 'Course', constructors: const {
-  '': const FunctionMirror(parameters: const {}, call: _Course__Constructor)
+  '': const FunctionMirror($call: _Course__Constructor)
 }, fields: const {
   'id': $$Course_fields_id,
   'name': $$Course_fields_name,