+++ test/scanners/GetPublicMembers_test.g.dart
@@ -6,34 +6,38 @@ part of GetVarialbesAndGettersTest_test;
 // Generator: MirrorsGenerator
 // **************************************************************************
 
-_ObjectWithMembers__Constructor(params) => new ObjectWithMembers();
+_ObjectWithMembers__Constructor([positionalParams, namedParams]) =>
+    new ObjectWithMembers();
 
 const $$ObjectWithMembers_fields_name = const DeclarationMirror(type: String);
 const $$ObjectWithMembers_fields_age = const DeclarationMirror(type: int);
 
-const ObjectWithMembersClassMirror =
-    const ClassMirror(name: 'ObjectWithMembers', constructors: const {
-  '': const FunctionMirror(
-      parameters: const {}, call: _ObjectWithMembers__Constructor)
-}, fields: const {
-  'name': $$ObjectWithMembers_fields_name,
-  'age': $$ObjectWithMembers_fields_age
-}, getters: const [
-  'name',
-  'age'
-], setters: const [
-  'name',
-  'age'
-]);
-_ExtendedObject__Constructor(params) => new ExtendedObject();
+const ObjectWithMembersClassMirror = const ClassMirror(
+    name: 'ObjectWithMembers',
+    constructors: const {
+      '': const FunctionMirror($call: _ObjectWithMembers__Constructor)
+    },
+    fields: const {
+      'name': $$ObjectWithMembers_fields_name,
+      'age': $$ObjectWithMembers_fields_age
+    },
+    getters: const [
+      'name',
+      'age'
+    ],
+    setters: const [
+      'name',
+      'age'
+    ]);
+_ExtendedObject__Constructor([positionalParams, namedParams]) =>
+    new ExtendedObject();
 
 const $$ExtendedObject_fields_myProp = const DeclarationMirror(type: String);
 
 const ExtendedObjectClassMirror = const ClassMirror(
     name: 'ExtendedObject',
     constructors: const {
-      '': const FunctionMirror(
-          parameters: const {}, call: _ExtendedObject__Constructor)
+      '': const FunctionMirror($call: _ExtendedObject__Constructor)
     },
     fields: const {
       'myProp': $$ExtendedObject_fields_myProp,