+++ test/scanners/Is_test.g.dart
@@ -6,16 +6,18 @@ part of Is_test;
 // Generator: MirrorsGenerator
 // **************************************************************************
 
-_ClassWithAnnotations__Constructor(params) => new ClassWithAnnotations();
+_ClassWithAnnotations__Constructor([positionalParams, namedParams]) =>
+    new ClassWithAnnotations();
 
-const ClassWithAnnotationsClassMirror =
-    const ClassMirror(name: 'ClassWithAnnotations', constructors: const {
-  '': const FunctionMirror(
-      parameters: const {}, call: _ClassWithAnnotations__Constructor)
-}, annotations: const [
-  const Annotation1(),
-  const Annotation1()
-]);
+const ClassWithAnnotationsClassMirror = const ClassMirror(
+    name: 'ClassWithAnnotations',
+    constructors: const {
+      '': const FunctionMirror($call: _ClassWithAnnotations__Constructor)
+    },
+    annotations: const [
+      const Annotation1(),
+      const Annotation1()
+    ]);
 
 // **************************************************************************
 // Generator: InitMirrorsGenerator